django_backend/requirements.txt
django
djangorestframework
json5
pandas
openpyxl
pycryptodome